requirements.txt
mysqlclient
pandas
pymssql
pytest
pytest-cov
pyyaml
sqlalchemy
